refactor(cli): extract remaining subcommands into commands/ modules
Move the last self-contained handlers out of main.rs: machine-load, fuzz,
ingest-svd/import-svd (svd.rs), codegen, coverage, tier1-matrix — one module
per command. All zero-coupling (use crate::* for the shared args/consts).

main.rs is now command dispatch + shared infra + arg structs only.
Pure move — build + clippy clean.
