+++ crates/cli/src/commands/codegen.rs
@@ -1,0 +1,57 @@
+// LabWired - Firmware Simulation Platform
+// Copyright (C) 2026 Andrii Shylenko
+//
+// This software is released under the MIT License.
+// See the LICENSE file in the project root for full license information.
+
+//! `labwired codegen` peripheral-descriptor codegen.
+
+use crate::*;
+
+pub(crate) fn run_codegen(args: CodegenArgs) -> ExitCode {
+    info!("Generating Rust code from IR: {:?}", args.input);
+
+    let file = match std::fs::File::open(&args.input) {
+        Ok(f) => f,
+        Err(e) => {
+            error!("Failed to open IR file: {}", e);
+            return ExitCode::from(EXIT_CONFIG_ERROR);
+        }
+    };
+
+    let device: labwired_ir::IrDevice = match serde_json::from_reader(file) {
+        Ok(d) => d,
+        Err(e) => {
+            error!("Failed to parse IR JSON: {}", e);
+            return ExitCode::from(EXIT_CONFIG_ERROR);
+        }
+    };
+
+    let mut output_code = String::new();
+    output_code.push_str("// Generated by LabWired Codegen\n");
+    output_code.push_str("#![allow(non_camel_case_types)]\n");
+    output_code.push_str("#![allow(non_snake_case)]\n");
+    // output_code.push_str("use labwired_core::Peripheral;\n"); // Not strictly needed yet as we generate structs
+    // output_code.push_str("use labwired_core::SimResult;\n\n");
+
+    for (name, peripheral) in &device.peripherals {
+        match labwired_codegen::PeripheralGenerator::generate(peripheral) {
+            Ok(code) => {
+                output_code.push_str(&code);
+                output_code.push_str("\n\n");
+            }
+            Err(e) => {
+                error!("Failed to generate code for peripheral {}: {}", name, e);
+                return ExitCode::from(EXIT_CONFIG_ERROR);
+            }
+        }
+    }
+
+    if let Err(e) = std::fs::write(&args.output, output_code) {
+        error!("Failed to write output file: {}", e);
+        return ExitCode::from(EXIT_CONFIG_ERROR);
+    }
+
+    info!("Successfully wrote Rust code to {:?}", args.output);
+    ExitCode::from(EXIT_PASS)
+}
