+++ crates/cli/src/commands/mod.rs
@@ -1,6 +1,12 @@
 //! CLI subcommand handlers, split out of `main.rs`.
 
 pub mod asset;
+pub mod codegen;
+pub mod coverage;
+pub mod fuzz;
+pub mod machine;
 pub mod run;
 pub mod snapshot;
+pub mod svd;
 pub mod test;
+pub mod tier1;
