+++ crates/cli/src/main.rs
@@ -843,9 +843,9 @@ fn main() -> ExitCode {
         Some(Commands::Asset(args)) => run_asset(args),
         Some(Commands::Run(args)) => commands::run::run_firmware(args),
         Some(Commands::Snapshot(args)) => commands::snapshot::run_snapshot(args),
-        Some(Commands::Coverage(args)) => run_coverage(args),
-        Some(Commands::Tier1Matrix(args)) => run_tier1_matrix(args),
-        Some(Commands::Fuzz(args)) => run_fuzz(args),
+        Some(Commands::Coverage(args)) => commands::coverage::run_coverage(args),
+        Some(Commands::Tier1Matrix(args)) => commands::tier1::run_tier1_matrix(args),
+        Some(Commands::Fuzz(args)) => commands::fuzz::run_fuzz(args),
         None => commands::run::run_interactive(cli),
     }
 }
@@ -1215,228 +1215,10 @@ fn run_two_c3_wifi(
     ExitCode::SUCCESS
 }
 
-fn run_coverage(args: CoverageArgs) -> ExitCode {
-    if let Some(p) = &args.svd {
-        std::env::set_var("LABWIRED_ESP32S3_SVD", p);
-    }
-    match coverage::run() {
-        Some((matrix, text)) => {
-            print!("{text}");
-            if let Some(out) = &args.json_out {
-                let json = serde_json::to_string_pretty(&matrix).expect("serialize matrix");
-                std::fs::write(out, &json).expect("write json");
-                eprintln!("wrote {}", out.display());
-            }
-            ExitCode::SUCCESS
-        }
-        None => {
-            eprintln!(
-                "error: ESP32-S3 SVD not found; set --svd or LABWIRED_ESP32S3_SVD, \
-                 or install the espressif32 PlatformIO platform"
-            );
-            ExitCode::from(EXIT_CONFIG_ERROR)
-        }
-    }
-}
-
-fn run_fuzz(args: FuzzArgs) -> ExitCode {
-    use labwired_fuzz::{fuzz, fuzz_collect, Contract, Target, Verdict};
-
-    let contract = Contract {
-        input_len: args.input_len_addr,
-        input_data: args.input_data_addr,
-        verdict: args.verdict_addr,
-        done_magic: args.done_magic,
-        fault_magic: args.fault_magic,
-    };
-
-    // Seeds: parse `--seed-input` hex bytes; empty means the engine self-seeds.
-    let mut seeds: Vec<Vec<u8>> = Vec::new();
-    for s in &args.seed_input {
-        let t = s.trim_start_matches("0x");
-        if t.len() % 2 != 0 {
-            eprintln!("error: --seed-input `{s}` must be an even number of hex digits");
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-        let mut bytes = Vec::with_capacity(t.len() / 2);
-        for i in (0..t.len()).step_by(2) {
-            match u8::from_str_radix(&t[i..i + 2], 16) {
-                Ok(b) => bytes.push(b),
-                Err(e) => {
-                    eprintln!("error: --seed-input `{s}`: {e}");
-                    return ExitCode::from(EXIT_CONFIG_ERROR);
-                }
-            }
-        }
-        seeds.push(bytes);
-    }
-
-    let target = match Target::from_elf(
-        &args.chip,
-        &args.system,
-        &args.firmware,
-        contract,
-        args.max_steps,
-    ) {
-        Ok(t) => t,
-        Err(e) => {
-            eprintln!("error: {e:#}");
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let engine = if cfg!(feature = "fuzz-libafl") {
-        "LibAFL"
-    } else {
-        "built-in"
-    };
-    eprintln!(
-        "fuzzing {} with the {engine} engine (max_iters={}, seed={:#x}) ...",
-        args.firmware.display(),
-        args.max_iters,
-        args.seed
-    );
-
-    // Collect-N mode gathers distinct crashes (feeds HIL-confirm); default mode
-    // stops at the first crash.
-    let crashes: Vec<Vec<u8>> = if let Some(n) = args.collect {
-        fuzz_collect(&target, seeds, args.max_iters, args.seed, n)
-    } else {
-        match fuzz(&target, seeds, args.max_iters, args.seed) {
-            r @ labwired_fuzz::FuzzReport { crash: None, .. } => {
-                println!(
-                    "no crash in {} iters (corpus {}, {} edges)",
-                    r.iterations, r.corpus_size, r.edges_hit
-                );
-                return ExitCode::SUCCESS;
-            }
-            labwired_fuzz::FuzzReport {
-                crash: Some(c),
-                iterations,
-                corpus_size,
-                edges_hit,
-            } => {
-                println!(
-                    "CRASH in {iterations} iters (corpus {corpus_size}, {edges_hit} edges): {:02X?}",
-                    c
-                );
-                vec![c]
-            }
-        }
-    };
-
-    if crashes.is_empty() {
-        println!("no crash found in {} iters", args.max_iters);
-        return ExitCode::SUCCESS;
-    }
-
-    if args.collect.is_some() {
-        println!("found {} distinct crash(es):", crashes.len());
-        for c in &crashes {
-            println!("  {c:02X?}");
-        }
-    }
-
-    // Reproduce + report the first crash's verdict for clarity.
-    let mut cov = labwired_fuzz::CovMap::new();
-    let verdict = target.run(&crashes[0], &mut cov);
-    let label = match verdict {
-        Verdict::Crash => "crash (fault/panic marker)",
-        Verdict::Hang => "hang (step budget exhausted)",
-        Verdict::Clean => "clean (non-deterministic?)",
-    };
-    eprintln!("first crash reproduces as: {label}");
-
-    if let Some(out) = &args.crashes_out {
-        match serde_json::to_string_pretty(&crashes) {
-            Ok(json) => {
-                if let Err(e) = std::fs::write(out, json) {
-                    eprintln!("error: write {}: {e}", out.display());
-                    return ExitCode::FAILURE;
-                }
-                eprintln!(
-                    "wrote {} crash input(s) to {}",
-                    crashes.len(),
-                    out.display()
-                );
-            }
-            Err(e) => {
-                eprintln!("error: serialize crashes: {e}");
-                return ExitCode::FAILURE;
-            }
-        }
-    }
-
-    // A crash is a finding — non-zero exit so CI fails the build.
-    ExitCode::FAILURE
-}
-
-fn run_tier1_matrix(args: Tier1MatrixArgs) -> ExitCode {
-    let self_bin = match std::env::current_exe() {
-        Ok(p) => p,
-        Err(e) => {
-            eprintln!("error: cannot resolve current executable: {e}");
-            return ExitCode::FAILURE;
-        }
-    };
-    match labwired_cli::tier1::run_all(&self_bin) {
-        Ok((mut matrix, skipped)) => {
-            for chip in &skipped {
-                eprintln!("SKIP: {chip} (fixture not present)");
-            }
-            // --run-url given but nothing was actually exercised → vacuous green
-            // is not permitted; fail loudly so CI notices the misconfiguration.
-            // (Skipped targets still emit unrecorded rows, so key on the count
-            // of EXERCISED chips, not on matrix emptiness.)
-            if args.run_url.is_some() && matrix.0.len() == skipped.len() {
-                eprintln!("error: --run-url given but no fixtures were exercised");
-                return ExitCode::FAILURE;
-            }
-            if let Some(url) = &args.run_url {
-                use labwired_cli::tier1::CellStatus;
-                for row in matrix.0.values_mut() {
-                    for cell in row.values_mut() {
-                        if cell.status != CellStatus::Unrecorded && cell.status != CellStatus::Na {
-                            cell.run_url = Some(url.clone());
-                        }
-                    }
-                }
-            }
-            // Text grid for humans.
-            for (chip, row) in &matrix.0 {
-                let cells: Vec<String> = row
-                    .iter()
-                    .map(|(class, cell)| format!("{class}={}", cell.status.as_str()))
-                    .collect();
-                println!("{chip}: {}", cells.join(" "));
-            }
-            if let Some(out) = &args.json_out {
-                let json = match serde_json::to_string_pretty(&matrix) {
-                    Ok(j) => j,
-                    Err(e) => {
-                        eprintln!("error: failed to serialize tier1 matrix: {e}");
-                        return ExitCode::FAILURE;
-                    }
-                };
-                if let Err(e) = std::fs::write(out, json.as_bytes()) {
-                    eprintln!("error: failed to write {}: {e}", out.display());
-                    return ExitCode::FAILURE;
-                }
-                eprintln!("wrote {}", out.display());
-            }
-            ExitCode::SUCCESS
-        }
-        Err(e) => {
-            eprintln!("tier1-matrix failed: {e}");
-            ExitCode::FAILURE
-        }
-    }
-}
-
 fn run_asset(args: AssetArgs) -> ExitCode {
     match args.command {
-        AssetCommands::ImportSvd(a) => run_import_svd(a),
-        AssetCommands::Codegen(a) => run_codegen(a),
+        AssetCommands::ImportSvd(a) => commands::svd::run_import_svd(a),
+        AssetCommands::Codegen(a) => commands::codegen::run_codegen(a),
         AssetCommands::Init(a) => commands::asset::run_asset_init(a),
         AssetCommands::AddPeripheral(a) => commands::asset::run_asset_add_peripheral(a),
         AssetCommands::Validate(a) => asset_validation::run_validate(a),
@@ -1444,7 +1226,7 @@ fn run_asset(args: AssetArgs) -> ExitCode {
         AssetCommands::Create(a) => commands::asset::run_asset_create(a),
         AssetCommands::Verify(a) => commands::asset::run_asset_verify(a),
         AssetCommands::ValidateComponent(a) => component_validation::run_validate_component(a),
-        AssetCommands::IngestSvd(a) => run_ingest_svd(a),
+        AssetCommands::IngestSvd(a) => commands::svd::run_ingest_svd(a),
     }
 }
 
@@ -1485,430 +1267,9 @@ pub(crate) fn resolve_chip_descriptor_path(chip: &str) -> Option<PathBuf> {
     None
 }
 
-fn run_codegen(args: CodegenArgs) -> ExitCode {
-    info!("Generating Rust code from IR: {:?}", args.input);
-
-    let file = match std::fs::File::open(&args.input) {
-        Ok(f) => f,
-        Err(e) => {
-            error!("Failed to open IR file: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let device: labwired_ir::IrDevice = match serde_json::from_reader(file) {
-        Ok(d) => d,
-        Err(e) => {
-            error!("Failed to parse IR JSON: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let mut output_code = String::new();
-    output_code.push_str("// Generated by LabWired Codegen\n");
-    output_code.push_str("#![allow(non_camel_case_types)]\n");
-    output_code.push_str("#![allow(non_snake_case)]\n");
-    // output_code.push_str("use labwired_core::Peripheral;\n"); // Not strictly needed yet as we generate structs
-    // output_code.push_str("use labwired_core::SimResult;\n\n");
-
-    for (name, peripheral) in &device.peripherals {
-        match labwired_codegen::PeripheralGenerator::generate(peripheral) {
-            Ok(code) => {
-                output_code.push_str(&code);
-                output_code.push_str("\n\n");
-            }
-            Err(e) => {
-                error!("Failed to generate code for peripheral {}: {}", name, e);
-                return ExitCode::from(EXIT_CONFIG_ERROR);
-            }
-        }
-    }
-
-    if let Err(e) = std::fs::write(&args.output, output_code) {
-        error!("Failed to write output file: {}", e);
-        return ExitCode::from(EXIT_CONFIG_ERROR);
-    }
-
-    info!("Successfully wrote Rust code to {:?}", args.output);
-    ExitCode::from(EXIT_PASS)
-}
-
-fn run_import_svd(args: ImportSvdArgs) -> ExitCode {
-    info!("Importing SVD from {:?}", args.input);
-
-    let xml = match std::fs::read_to_string(&args.input) {
-        Ok(s) => s,
-        Err(e) => {
-            error!("Failed to read SVD file: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let svd = match svd_parser::parse(&xml) {
-        Ok(s) => s,
-        Err(e) => {
-            error!("Failed to parse SVD XML: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let mut device = match labwired_ir::IrDevice::from_svd(&svd) {
-        Ok(d) => d,
-        Err(e) => {
-            error!("Failed to convert to Strict IR: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    if let (Some(base), Some(size)) = (args.flash_base, &args.flash_size) {
-        device.memory_regions.insert(
-            "FLASH".to_string(),
-            labwired_ir::IrMemoryRegion {
-                name: "FLASH".to_string(),
-                base: base as u64,
-                size: labwired_config::parse_size(size)
-                    .map_err(|e| {
-                        error!("Invalid flash size '{}': {}", size, e);
-                    })
-                    .unwrap_or(0),
-            },
-        );
-    }
-
-    if let (Some(base), Some(size)) = (args.ram_base, &args.ram_size) {
-        device.memory_regions.insert(
-            "RAM".to_string(),
-            labwired_ir::IrMemoryRegion {
-                name: "RAM".to_string(),
-                base: base as u64,
-                size: labwired_config::parse_size(size)
-                    .map_err(|e| {
-                        error!("Invalid ram size '{}': {}", size, e);
-                    })
-                    .unwrap_or(0),
-            },
-        );
-    }
-
-    let file = match std::fs::File::create(&args.output) {
-        Ok(f) => f,
-        Err(e) => {
-            error!("Failed to create output file: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    if let Err(e) = serde_json::to_writer_pretty(file, &device) {
-        error!("Failed to write JSON: {}", e);
-        return ExitCode::from(EXIT_CONFIG_ERROR);
-    }
-
-    info!("Successfully wrote Strict IR to {:?}", args.output);
-    ExitCode::from(EXIT_PASS)
-}
-
-fn run_ingest_svd(args: IngestSvdArgs) -> ExitCode {
-    // Keep stdout pure JSON in --json mode (the MCP agent surface parses it);
-    // the progress line is only useful for the human table mode.
-    if !args.json {
-        info!(
-            "Ingesting SVD {:?} -> declarative descriptors in {:?}",
-            args.input, args.output_dir
-        );
-    }
-
-    let xml = match std::fs::read_to_string(&args.input) {
-        Ok(s) => s,
-        Err(e) => {
-            error!("Failed to read SVD file: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-    let device = match svd_parser::parse(&xml) {
-        Ok(d) => d,
-        Err(e) => {
-            error!("Failed to parse SVD XML: {}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-    if let Err(e) = std::fs::create_dir_all(&args.output_dir) {
-        error!("Failed to create output directory: {}", e);
-        return ExitCode::from(EXIT_CONFIG_ERROR);
-    }
-
-    let filter: Option<Vec<String>> = args.filter.as_ref().map(|s| {
-        s.split(',')
-            .map(|t| t.trim().to_string())
-            .filter(|t| !t.is_empty())
-            .collect()
-    });
-
-    let mut summary: Vec<serde_json::Value> = Vec::new();
-    let mut errors = 0usize;
-    for peripheral in &device.peripherals {
-        if let Some(ref f) = filter {
-            if !f.iter().any(|n| n.eq_ignore_ascii_case(&peripheral.name)) {
-                continue;
-            }
-        }
-        match svd_ingestor::process_peripheral(&device, peripheral) {
-            Ok(desc) => {
-                if let Err(e) = svd_ingestor::save_descriptor(&desc, &args.output_dir) {
-                    error!("Failed to save descriptor for {}: {}", peripheral.name, e);
-                    errors += 1;
-                    continue;
-                }
-                let path = args
-                    .output_dir
-                    .join(format!("{}.yaml", desc.peripheral.to_lowercase()));
-                summary.push(serde_json::json!({
-                    "name": desc.peripheral,
-                    "descriptor_path": path.to_string_lossy(),
-                    "register_count": desc.registers.len(),
-                    "base_address": format!("0x{:08X}", peripheral.base_address),
-                }));
-            }
-            Err(e) => {
-                error!("Failed to process peripheral {}: {}", peripheral.name, e);
-                errors += 1;
-            }
-        }
-    }
-
-    if args.json {
-        let out = serde_json::json!({
-            "output_dir": args.output_dir.to_string_lossy(),
-            "peripheral_count": summary.len(),
-            "peripherals": summary,
-        });
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&out).unwrap_or_else(|_| "{}".into())
-        );
-    } else {
-        println!(
-            "Ingested {} peripheral(s) into {}:",
-            summary.len(),
-            args.output_dir.display()
-        );
-        for p in &summary {
-            println!(
-                "  {:<16} {:>4} registers -> {}",
-                p["name"].as_str().unwrap_or("?"),
-                p["register_count"].as_u64().unwrap_or(0),
-                p["descriptor_path"].as_str().unwrap_or("?")
-            );
-        }
-    }
-
-    if summary.is_empty() {
-        error!("No peripherals ingested (check --filter against the SVD's peripheral names)");
-        return ExitCode::from(EXIT_CONFIG_ERROR);
-    }
-    if errors > 0 {
-        return ExitCode::from(EXIT_CONFIG_ERROR);
-    }
-    ExitCode::from(EXIT_PASS)
-}
-
 fn run_machine(args: MachineArgs) -> ExitCode {
     match args.command {
-        MachineCommands::Load(load_args) => run_machine_load(load_args),
-    }
-}
-
-fn run_machine_load(args: LoadArgs) -> ExitCode {
-    info!("Loading machine from snapshot: {:?}", args.snapshot);
-
-    let f = match std::fs::File::open(&args.snapshot) {
-        Ok(f) => f,
-        Err(e) => {
-            error!("Failed to open snapshot {:?}: {}", args.snapshot, e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let snapshot_data: Snapshot = match serde_json::from_reader(f) {
-        Ok(s) => s,
-        Err(e) => {
-            error!("Failed to parse snapshot {:?}: {}", args.snapshot, e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let (cpu_snapshot, peripherals_snapshot, config) = match snapshot_data {
-        Snapshot::Interactive {
-            cpu,
-            peripherals,
-            config,
-            ..
-        } => (cpu, peripherals, config),
-        _ => {
-            error!("Unsupported snapshot type for loading");
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    // Reconstruct bus
-    let mut bus = match labwired_core::system::builder::build_system_bus(config.system.as_deref()) {
-        Ok(bus) => bus,
-        Err(e) => {
-            error!("Failed to reconstruct bus: {:#}", e);
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    // Load original firmware (required for memory content that isn't in snapshot yet)
-    // Note: Our snapshot currently doesn't include full RAM/Flash dumps to keep it small.
-    // So we MUST load the firmware first.
-    let program = match labwired_loader::load_elf(&config.firmware) {
-        Ok(p) => p,
-        Err(e) => {
-            error!(
-                "Failed to load original firmware {:?}: {:#}",
-                config.firmware, e
-            );
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let arch = match program.arch {
-        labwired_core::Arch::Arm => labwired_config::Arch::Arm,
-        labwired_core::Arch::RiscV => labwired_config::Arch::RiscV,
-        labwired_core::Arch::XtensaLx7 => labwired_config::Arch::Xtensa,
-        _ => {
-            error!("Unknown architecture in firmware");
-            return ExitCode::from(EXIT_CONFIG_ERROR);
-        }
-    };
-
-    let metrics = Arc::new(labwired_core::metrics::PerformanceMetrics::new());
-
-    match arch {
-        labwired_config::Arch::Arm => {
-            let (cpu, _) = labwired_core::system::cortex_m::configure_cortex_m(&mut bus);
-            let mut machine = labwired_core::Machine::new(cpu, bus);
-            machine.observers.push(metrics.clone());
-            if let Err(e) = machine.load_firmware(&program) {
-                error!("Failed to load firmware: {}", e);
-                return ExitCode::from(EXIT_RUNTIME_ERROR);
-            }
-
-            // Apply snapshot
-            machine.cpu.apply_snapshot(&cpu_snapshot);
-            for ps in peripherals_snapshot {
-                if let Some(state) = ps.state {
-                    // Find peripheral by name and restore
-                    for p in &mut machine.bus.peripherals {
-                        if p.name == ps.name {
-                            if let Err(e) = p.dev.restore(state.clone()) {
-                                error!("Failed to restore peripheral {}: {}", p.name, e);
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-
-            info!("Resuming simulation (ARM)...");
-            let cli = Cli {
-                firmware: Some(config.firmware),
-                system: config.system,
-                snapshot: None,
-                breakpoint: vec![],
-                trace: args.trace,
-                max_steps: args.max_steps.unwrap_or(config.max_steps),
-                gdb: None,
-                command: None,
-                json: false,
-                vcd: None,
-            };
-            run_simulation_loop(&cli, &mut machine, &metrics);
-            ExitCode::from(EXIT_PASS)
-        }
-        labwired_config::Arch::RiscV => {
-            let cpu = labwired_core::system::riscv::configure_riscv(&mut bus);
-            let mut machine = labwired_core::Machine::new(cpu, bus);
-            machine.observers.push(metrics.clone());
-            if let Err(e) = machine.load_firmware(&program) {
-                error!("Failed to load firmware: {}", e);
-                return ExitCode::from(EXIT_RUNTIME_ERROR);
-            }
-
-            // Apply snapshot
-            machine.cpu.apply_snapshot(&cpu_snapshot);
-            for ps in peripherals_snapshot {
-                if let Some(state) = ps.state {
-                    for p in &mut machine.bus.peripherals {
-                        if p.name == ps.name {
-                            if let Err(e) = p.dev.restore(state.clone()) {
-                                error!("Failed to restore peripheral {}: {}", p.name, e);
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-
-            info!("Resuming simulation (RISC-V)...");
-            let cli = Cli {
-                firmware: Some(config.firmware),
-                system: config.system,
-                snapshot: None,
-                breakpoint: vec![],
-                trace: args.trace,
-                max_steps: args.max_steps.unwrap_or(config.max_steps),
-                gdb: None,
-                command: None,
-                json: false,
-                vcd: None,
-            };
-            run_simulation_loop(&cli, &mut machine, &metrics);
-            ExitCode::from(EXIT_PASS)
-        }
-        labwired_config::Arch::Xtensa => {
-            let cpu = labwired_core::system::xtensa::configure_xtensa(&mut bus);
-            let mut machine = labwired_core::Machine::new(cpu, bus);
-            machine.observers.push(metrics.clone());
-            if let Err(e) = machine.load_firmware(&program) {
-                error!("Failed to load firmware: {}", e);
-                return ExitCode::from(EXIT_RUNTIME_ERROR);
-            }
-
-            machine.cpu.apply_snapshot(&cpu_snapshot);
-            for ps in peripherals_snapshot {
-                if let Some(state) = ps.state {
-                    for p in &mut machine.bus.peripherals {
-                        if p.name == ps.name {
-                            if let Err(e) = p.dev.restore(state.clone()) {
-                                error!("Failed to restore peripheral {}: {}", p.name, e);
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-
-            info!("Resuming simulation (Xtensa)...");
-            let cli = Cli {
-                firmware: Some(config.firmware),
-                system: config.system,
-                snapshot: None,
-                breakpoint: vec![],
-                trace: args.trace,
-                max_steps: args.max_steps.unwrap_or(config.max_steps),
-                gdb: None,
-                command: None,
-                json: false,
-                vcd: None,
-            };
-            run_simulation_loop(&cli, &mut machine, &metrics);
-            ExitCode::from(EXIT_PASS)
-        }
-        _ => {
-            error!("Unsupported architecture for snapshot load");
-            ExitCode::from(EXIT_CONFIG_ERROR)
-        }
+        MachineCommands::Load(load_args) => commands::machine::run_machine_load(load_args),
     }
 }
